with completion

+++ app/src/main/java/com/example/labtest1/feeskeeper/myplaces3/MapsActivity.kt
@@ -146,7 +146,7 @@ private lateinit var wordViewModel: feeViewModel
 
       private fun gotoupdate(title: String) {
           val intent = Intent(this , addAct::class.java)
-         addAct.booze = title
+         addAct.booze = title.toInt()
           addAct.isupdate = true
           startActivity(intent)
       }
